requirements.txt
Django==4.2.7
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-telegram-bot==13.15
geopy==2.4.0
django-redis==5.4.0